requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
websockets==14.0
twilio==9.4.0
faster-whisper==1.2.1
edge-tts==7.0.0
httpx==0.28.0
numpy>=2.1.0
scipy>=1.14.0
torch>=2.5.0
pyyaml>=6.0.2
python-dotenv==1.0.1
audioop-lts==0.2.2
pydantic>=2.10.0
